+++ pkg/gatewayserver/io/ws/ws.go
@@ -31,6 +31,7 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	"go.thethings.network/lorawan-stack/v3/pkg/frequencyplans"
 	"go.thethings.network/lorawan-stack/v3/pkg/gatewayserver/io"
+	"go.thethings.network/lorawan-stack/v3/pkg/gatewayserver/scheduling"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	"go.thethings.network/lorawan-stack/v3/pkg/random"
 	"go.thethings.network/lorawan-stack/v3/pkg/ratelimit"
@@ -236,7 +237,6 @@ func (s *srv) handleTraffic(w http.ResponseWriter, r *http.Request) (err error) 
 	defer pingTicker.Stop()
 
 	ws.SetPingHandler(func(data string) error {
-		logger.Debug("Received ping from gateway, send pong")
 		select {
 		case <-ctx.Done():
 			return ctx.Err()
@@ -248,7 +248,6 @@ func (s *srv) handleTraffic(w http.ResponseWriter, r *http.Request) (err error) 
 	// Not all gateways support pongs to the server's pings.
 	ws.SetPongHandler(func(data string) error {
 		atomic.StoreInt64(&missedPongs, 0)
-		logger.Debug("Received pong from gateway")
 		return nil
 	})
 
@@ -304,12 +303,7 @@ func (s *srv) handleTraffic(w http.ResponseWriter, r *http.Request) (err error) 
 					return err
 				}
 			case down := <-conn.Down():
-				concentratorTime, ok := conn.TimeFromTimestampTime(down.GetScheduled().Timestamp)
-				if !ok {
-					logger.Warn("No clock synchronization")
-					continue
-				}
-				dnmsg, err := s.formatter.FromDownlink(ctx, *down, conn.BandID(), concentratorTime, time.Now())
+				dnmsg, err := s.formatter.FromDownlink(ctx, *down, conn.BandID(), scheduling.ConcentratorTime(down.GetScheduled().ConcentratorTimestamp), time.Now())
 				if err != nil {
 					logger.WithError(err).Warn("Failed to marshal downlink message")
 					continue